pyproject.toml
[tool.poetry]
name = "chatbotteste"
version = "0.1.0"
description = ""
authors = ["pedroaltobelli23 <pedroaltobelli673@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
